Rename web to web-nunjucks
