+++ README.md
@@ -14,7 +14,7 @@
 
     ```json  
     "generate:plop": "plop --plopfile ./node_modules/govuk-prototype-kit-plops/plopfile.js",
-    "generate:web": "./node_modules/govuk-prototype-kit-plops/web/bin/www",
+    "generate:web": "./node_modules/govuk-prototype-kit-plops/web-nunjucks/bin/www",
     ```
 
 3. From within the prototype kit run `npm run generate:plop` and then answer the CLI accordingly and for the web run `npm run generate:web`
