+++ package.json
@@ -30,13 +30,13 @@
   },
   "devDependencies": {
     "@dwp/eslint-config-base": "6.0.0",
-    "@jest/globals": "^29.2.2",
+    "@jest/globals": "29.2.2",
     "eslint": "8.26.0",
     "eslint-plugin-no-unsafe-regex": "1.0.0",
     "eslint-plugin-security": "1.5.0",
     "eslint-plugin-sonarjs": "0.16.0",
     "husky": "8.0.1",
-    "jest": "^29.2.2",
+    "jest": "29.2.2",
     "lint-staged": "13.0.3"
   },
   "lint-staged": {
