+++ web-nunjucks/views/guidance-page.njk
@@ -16,7 +16,7 @@
                 text: "Enter the page name"
             },
             hint: {
-                text: "This is used to creat the page URL and the file names"
+                text: "This is used to create the page URL and the file names"
             },
             id: "pageName",
             name: "pageName"
